+++ stwo_cairo_prover/crates/prover/src/adapter/builtins/test_artifacts/test_public_input.json
